+++ versions.sh
@@ -13,9 +13,11 @@ oc version | head -1
 # terraform
 terraform --version | head -1
 # kubectl
-echo -n "kubectl: " && kubectl version --short=true 2>/dev/null | head -1 | cut -d' ' -f3
-# awscli
+echo -n "kubectl: " && kubectl version -o json | jq '.clientVersion.gitVersion'
+#kubectl version --short=true 2>/dev/null | head -1 | cut -d' ' -f3
+# aws
 pip list 2>/dev/null|grep awscli
+echo -n "aws-iam-authenticator: " && aws-iam-authenticator version | jq '.Version'
 # gcloud
 echo -n "gcloud: " && gcloud --version |head -1 | cut -d' ' -f4
 # azure